modify  1. move package presentation pattern  2. implemented intent bridge
