+++ app/src/main/java/com/ama/algorithmmanagement/presentation/tryhistory/PagerIdeaFragment.kt
@@ -1,4 +1,4 @@
-package com.ama.algorithmmanagement.Fragment
+package com.ama.algorithmmanagement.presentation.tryhistory
 
 import android.app.Activity.RESULT_OK
 import android.app.AlertDialog
@@ -18,14 +18,13 @@ import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.core.content.FileProvider
 import androidx.lifecycle.ViewModelProvider
-import com.ama.algorithmmanagement.adapter.IdeaAdapter
 import com.ama.algorithmmanagement.application.AMAApplication
 import com.ama.algorithmmanagement.domain.base.BaseViewModelFactory
 import com.ama.algorithmmanagement.domain.base.KBaseFragment
 import com.ama.algorithmmanagement.R
 import com.ama.algorithmmanagement.data.repositories.RepositoryLocator
 import com.ama.algorithmmanagement.databinding.FragmentIdeaBinding
-import com.ama.algorithmmanagement.viewmodel.kDefault.MyIdeaInfoViewModel
+import com.ama.algorithmmanagement.presentation.tryhistory.adapter.IdeaAdapter
 import com.google.firebase.storage.FirebaseStorage
 import com.gun0912.tedpermission.PermissionListener
 import com.gun0912.tedpermission.TedPermission
