+++ app/src/main/java/com/ama/algorithmmanagement/presentation/tryhistory/TryHistoryViewModel.kt
@@ -1,4 +1,4 @@
-package com.ama.algorithmmanagement.viewmodel.kDefault
+package com.ama.algorithmmanagement.presentation.tryhistory
 
 import androidx.lifecycle.*
 import com.ama.algorithmmanagement.domain.base.BaseRepository
