+++ app/src/main/java/com/ama/algorithmmanagement/presentation/tryhistory/adapter/IdeaAdapter.kt
@@ -1,8 +1,8 @@
-package com.ama.algorithmmanagement.adapter
+package com.ama.algorithmmanagement.presentation.tryhistory.adapter
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.ama.algorithmmanagement.viewHolder.IdeaViewHolder
+import com.ama.algorithmmanagement.presentation.tryhistory.view_holder.IdeaViewHolder
 import com.ama.algorithmmanagement.data.model.IdeaInfo
 
 class IdeaAdapter : RecyclerView.Adapter<IdeaViewHolder>() {
