+++ app/src/main/java/com/ama/algorithmmanagement/utils/BindingAdapterUtils.kt
@@ -12,6 +12,11 @@ import com.ama.algorithmmanagement.Fragment.SolvedProblemViewPagerFragment
 import com.ama.algorithmmanagement.data.model.*
 import com.ama.algorithmmanagement.presentation.main.adapter.KRetryProblemsAdapter
 import com.ama.algorithmmanagement.presentation.main.adapter.KUserDateInfoAdapter
+import com.ama.algorithmmanagement.presentation.notip.adapter.NoTipProblemsAdapter
+import com.ama.algorithmmanagement.presentation.mytip.adapter.MyTipProblemsAdapter
+import com.ama.algorithmmanagement.presentation.tryfailed.adapter.TryFailedAdapter
+import com.ama.algorithmmanagement.presentation.tryhistory.adapter.IdeaAdapter
+import com.ama.algorithmmanagement.presentation.tryhistory.adapter.MyCommentAdapter
 import com.github.mikephil.charting.charts.HorizontalBarChart
 import com.github.mikephil.charting.charts.PieChart
 import com.github.mikephil.charting.components.XAxis
