+++ app/src/main/res/layout/activity_my_child_comment.xml
@@ -2,12 +2,12 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".activity.kDefault.MyChildCommentActivity">
+    tools:context=".presentation.tryhistory.MyChildCommentActivity">
 
     <data>
         <variable
             name="viewModel"
-            type="com.ama.algorithmmanagement.viewmodel.kDefault.MyChildCommentViewModel" />
+            type="com.ama.algorithmmanagement.presentation.tryhistory.MyChildCommentViewModel" />
     </data>
