+++ app/src/main/res/layout/fragment_idea.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.ama.algorithmmanagement.viewmodel.kDefault.MyIdeaInfoViewModel" />
+            type="com.ama.algorithmmanagement.presentation.tryhistory.MyIdeaInfoViewModel" />
     </data>
 
     <FrameLayout
